add comments and fix small bugs
