+++ src/libs/Navigation/helpers/dynamicRoutesUtils/getPathWithoutDynamicSuffix.ts
@@ -1,29 +1,6 @@
 import type {Route} from '@src/ROUTES';
-import {DYNAMIC_ROUTES} from '@src/ROUTES';
+import getDynamicRouteQueryParams from './getDynamicRouteQueryParams';
 import splitPathAndQuery from './splitPathAndQuery';
-
-/**
- * Returns the query parameter names that belong to a dynamic route suffix and should be
- * stripped when removing that suffix from the path. Looks up the matching DYNAMIC_ROUTES
- * config and returns its `queryParams` array if defined.
- *
- * @param dynamicSuffix - The dynamic route path segment (e.g., 'country')
- * @returns The list of query param keys to strip, or undefined if no match or no queryParams config
- *
- * @private - Internal helper. Do not export or use outside this file.
- */
-function getQueryParamsToStrip(dynamicSuffix: string): readonly string[] | undefined {
-    const keys = Object.keys(DYNAMIC_ROUTES) as Array<keyof typeof DYNAMIC_ROUTES>;
-    const match = keys.find((key) => DYNAMIC_ROUTES[key].path === dynamicSuffix);
-    if (!match) {
-        return undefined;
-    }
-    const config = DYNAMIC_ROUTES[match];
-    if (!('queryParams' in config)) {
-        return undefined;
-    }
-    return config.queryParams;
-}
 
 /**
  * Returns the path without a dynamic route suffix, stripping suffix-specific query parameters
@@ -41,7 +18,7 @@ function getPathWithoutDynamicSuffix(fullPath: string, dynamicSuffix: string): R
         return '';
     }
 
-    const paramsToStrip = getQueryParamsToStrip(dynamicSuffix);
+    const paramsToStrip = getDynamicRouteQueryParams(dynamicSuffix);
     let filteredQuery = query;
     if (paramsToStrip?.length && query) {
         const params = new URLSearchParams(query);
